// frontend/src/api/generated/jobs/jobs.ts
export type getUtilityJobApiV1JobsJobIdGetResponse200 = {
  data: UtilityJobRead;
  status: 200;
};

export type getUtilityJobApiV1JobsJobIdGetResponse422 = {
  data: HTTPValidationError;
  status: 422;
};

export type getUtilityJobApiV1JobsJobIdGetResponseSuccess =
  getUtilityJobApiV1JobsJobIdGetResponse200 & {
    headers: Headers;
  };
export type getUtilityJobApiV1JobsJobIdGetResponseError =
  getUtilityJobApiV1JobsJobIdGetResponse422 & {
    headers: Headers;
  };

export type getUtilityJobApiV1JobsJobIdGetResponse =
  | getUtilityJobApiV1JobsJobIdGetResponseSuccess
  | getUtilityJobApiV1JobsJobIdGetResponseError;